final-pjt-back/requirements.txt
asgiref==3.5.2
Django==3.2.13
django-cors-headers==4.0.0
djangorestframework==3.14.0
pytz==2022.6
sqlparse==0.4.3
